#7 add test of internal pkg and shared pkg
